+++ components/common/index.js
@@ -1,0 +1,3 @@
+import SEO from './SEO';
+
+export { SEO };
