+++ components/home/index.js
@@ -1,0 +1,5 @@
+import Header from './Header';
+import HeaderIcon from './HeaderIcon';
+import Welcome from './Welcome';
+
+export { Header, HeaderIcon, Welcome };
