+++ pages/index.js
@@ -1,6 +1,5 @@
-import SEO from '../components/SEO/SEO';
-import Header from '../components/Header/Header';
-import Welcome from '../components/Welcome/Welcome';
+import { SEO } from '../components/common';
+import { Header, Welcome } from '../components/home';
 import { Container } from '@chakra-ui/react';
 import React from 'react';
 
@@ -8,7 +7,7 @@ export default function Home() {
   return (
     <React.Fragment>
       <SEO />
-      <Container maxW='1400px' centerContent>
+      <Container p={0} maxW='1600px' centerContent>
         <Header />
         <Welcome />
       </Container>
